+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to the Aptos TypeScript SDK will be captured in this file. T
 # Unreleased
 
 - TypeTag parsing now support references, uppercase types, and more complete error handling
+- Allow simple string inputs as type arguments in move scripts
 
 # 1.16.0 (2024-05-22)